Comment stats and likes; Loading user like type in service layer

+++ src/main/java/com/kereq/main/service/PostStatisticsService.java
@@ -1,9 +1,10 @@
 package com.kereq.main.service;
 
 import com.kereq.common.error.RepositoryError;
-import com.kereq.main.constant.LikeType;
+import com.kereq.main.entity.PostLikeData;
 import com.kereq.main.entity.PostStatisticsData;
 import com.kereq.main.exception.ApplicationException;
+import com.kereq.main.repository.PostLikeRepository;
 import com.kereq.main.repository.PostRepository;
 import com.kereq.main.repository.PostStatisticsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,11 +17,21 @@ public class PostStatisticsService {
     private PostRepository postRepository;
 
     @Autowired
+    private PostLikeRepository postLikeRepository;
+
+    @Autowired
     private PostStatisticsRepository postStatisticsRepository;
 
-    public PostStatisticsData getStatistics(Long postId) {
-        return postStatisticsRepository.findById(postId)
-                .orElseThrow(() -> new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND));
+    public PostStatisticsData getStatistics(Long userId, Long postId) {
+        PostStatisticsData postStatistics = postStatisticsRepository.findByPostId(postId);
+        if (postStatistics == null) {
+            throw new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND);
+        }
+        PostLikeData postLikeData = postLikeRepository.findByUserIdAndPostId(userId, postId);
+        if (postLikeData != null) {
+            postStatistics.setUserLikeType(postLikeData.getType());
+        }
+        return postStatistics;
     }
 
     public PostStatisticsData initialize(Long postId) {
@@ -36,57 +47,73 @@ public class PostStatisticsService {
     }
 
     public void remove(Long postId) {
-        PostStatisticsData postStatistics = postStatisticsRepository.findById(postId)
-                .orElseThrow(() -> new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND));
+        PostStatisticsData postStatistics = postStatisticsRepository.findByPostId(postId);
+        if (postStatistics == null) {
+            throw new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND);
+        }
         postStatisticsRepository.delete(postStatistics);
     }
 
     public PostStatisticsData convertLike(Long postId, boolean toLike) {
-        PostStatisticsData postStatistics = postStatisticsRepository.findById(postId)
-                .orElseThrow(() -> new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND));
+        PostStatisticsData postStatistics = postStatisticsRepository.findByPostId(postId);
+        if (postStatistics == null) {
+            throw new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND);
+        }
         postStatistics.setLikesCount(postStatistics.getLikesCount() + (toLike ? 1 : -1));
         postStatistics.setDislikesCount(postStatistics.getDislikesCount() + (toLike ? -1 : 1));
         return postStatistics;
     }
 
     public PostStatisticsData addLike(Long postId) {
-        PostStatisticsData postStatistics = postStatisticsRepository.findById(postId)
-                .orElseThrow(() -> new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND));
+        PostStatisticsData postStatistics = postStatisticsRepository.findByPostId(postId);
+        if (postStatistics == null) {
+            throw new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND);
+        }
         postStatistics.setLikesCount(postStatistics.getLikesCount() + 1);
         return postStatistics;
     }
 
     public PostStatisticsData removeLike(Long postId) {
-        PostStatisticsData postStatistics = postStatisticsRepository.findById(postId)
-                .orElseThrow(() -> new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND));
+        PostStatisticsData postStatistics = postStatisticsRepository.findByPostId(postId);
+        if (postStatistics == null) {
+            throw new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND);
+        }
         postStatistics.setLikesCount(postStatistics.getLikesCount() - 1);
         return postStatistics;
     }
 
     public PostStatisticsData addDislike(Long postId) {
-        PostStatisticsData postStatistics = postStatisticsRepository.findById(postId)
-                .orElseThrow(() -> new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND));
+        PostStatisticsData postStatistics = postStatisticsRepository.findByPostId(postId);
+        if (postStatistics == null) {
+            throw new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND);
+        }
         postStatistics.setDislikesCount(postStatistics.getDislikesCount() + 1);
         return postStatistics;
     }
 
     public PostStatisticsData removeDislike(Long postId) {
-        PostStatisticsData postStatistics = postStatisticsRepository.findById(postId)
-                .orElseThrow(() -> new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND));
+        PostStatisticsData postStatistics = postStatisticsRepository.findByPostId(postId);
+        if (postStatistics == null) {
+            throw new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND);
+        }
         postStatistics.setDislikesCount(postStatistics.getDislikesCount() - 1);
         return postStatistics;
     }
 
     public PostStatisticsData addComment(Long postId) {
-        PostStatisticsData postStatistics = postStatisticsRepository.findById(postId)
-                .orElseThrow(() -> new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND));
+        PostStatisticsData postStatistics = postStatisticsRepository.findByPostId(postId);
+        if (postStatistics == null) {
+            throw new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND);
+        }
         postStatistics.setCommentsCount(postStatistics.getCommentsCount() + 1);
         return postStatistics;
     }
 
     public PostStatisticsData removeComment(Long postId) {
-        PostStatisticsData postStatistics = postStatisticsRepository.findById(postId)
-                .orElseThrow(() -> new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND));
+        PostStatisticsData postStatistics = postStatisticsRepository.findByPostId(postId);
+        if (postStatistics == null) {
+            throw new ApplicationException(RepositoryError.RESOURCE_NOT_FOUND);
+        }
         postStatistics.setCommentsCount(postStatistics.getCommentsCount() - 1);
         return postStatistics;
     }
